:construction:

+++ brew-cask.sh
@@ -1,12 +1,50 @@
 #!/bin/bash
 
+is_cask_exists() {
+  brew cask list | grep "$1"
+  return $?
+}
+
 echo "Install GUI Apps"
-brew cask install appcleaner
-brew cask install cyberduck
-brew cask install docker
-brew cask install dropbox
-brew cask install firefox
-brew cask install google-chrome
-brew cask install google-japanese-ime
-brew cask install iterm2
-brew cask install spotify
+
+if ! is_cask_exists "appcleaner"; then
+  brew cask install appcleaner
+fi
+
+if ! is_cask_exists "cyberduck"; then
+  brew cask install cyberduck
+fi
+
+if ! is_cask_exists "docker"; then
+  brew cask install docker
+fi
+
+if ! is_cask_exists "dropbox"; then
+  brew cask install dropbox
+fi
+
+if ! is_cask_exists "firefox"; then
+  brew cask install firefox
+fi
+
+if ! is_cask_exists "google-chrome"; then
+  brew cask install google-chrome
+fi
+
+if ! is_cask_exists "google-chrome-canary"; then
+  brew cask install google-chrome-canary
+fi
+
+if ! is_cask_exists "google-japanese-ime"; then
+  brew cask install google-japanese-ime
+fi
+
+if ! is_cask_exists "iterm2"; then
+  brew cask install iterm2
+fi
+
+
+if ! is_cask_exists "spotify"; then
+  brew cask install spotify
+fi
+
